+++ src/components/MainHeader/styles.ts
@@ -1,13 +1,12 @@
 import styled from "styled-components/native";
 import Animated from "react-native-reanimated";
 import Icon from "@expo/vector-icons/Feather";
-import { Dimensions, TextInput, View } from "react-native";
+import { Dimensions, TextInput } from "react-native";
 
-const { width, height } = Dimensions.get("window");
+const { width } = Dimensions.get("window");
 
 export const AnimatedIcon = Animated.createAnimatedComponent(Icon);
 const AnimatedInputSearch = Animated.createAnimatedComponent(TextInput)
-const AnimatedContainer = Animated.createAnimatedComponent(View);
 
 export const Container = styled.View`
   flex-direction: row;
@@ -15,13 +14,13 @@ export const Container = styled.View`
   justify-content: space-between;
 
   height: 80px;
-  padding: 0 24px;
+  padding: 0 16px;
   margin-top: 32px;
   background: transparent;
 `;
 
 export const InputSearch = styled(AnimatedInputSearch)`
-  height: 48px;
+  height: 56px;
   width: ${width * 0.7}px;
   border-radius: 12px;
   background: ${({ theme }) => theme.colors.secundary};