+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.2.1
+- [x] Fixed some animated components: **Main Footer**, **Main Header**, **Menu**.
+- [x] Fixed animated screen: **Tasks**.
+- [x] Fixed animated component when pressed hardware back button.
+
+### 0.2.0
+
+**Main Footer**
+- [x] Create main footer.
+
 ### 0.1.0
 
 **Main Header**